+++ lib/screens/opening_screen.dart
@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:ywsos2021_app/screens/home_screen.dart';
 import 'package:ywsos2021_app/screens/login_screen.dart';
 import 'package:ywsos2021_app/screens/register_screen.dart';
+import 'package:ywsos2021_app/widgets/opening_button.dart';
 
 class OpeningScreen extends StatelessWidget {
   const OpeningScreen({Key? key}) : super(key: key);
@@ -101,32 +102,6 @@ class OpeningScreen extends StatelessWidget {
           ],
         ),
       ),
-    );
-  }
-}
-
-class OpeningButton extends StatelessWidget {
-  final VoidCallback onPressed;
-  final Widget child;
-
-  const OpeningButton({
-    Key? key,
-    required this.onPressed,
-    required this.child,
-  }) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return ElevatedButton(
-      onPressed: onPressed,
-      child: child,
-      style: ElevatedButton.styleFrom(
-          shape:
-              RoundedRectangleBorder(borderRadius: BorderRadius.circular(15)),
-          primary: Color(0xFF97AC94),
-          padding: EdgeInsets.only(
-              left: 41, right: 41, top: 14, bottom: 20) // foreground
-          ),
     );
   }
 }
