+++ lib/widgets/dot_indicator.dart
@@ -1,0 +1,38 @@
+import 'package:carousel_slider/carousel_slider.dart';
+import 'package:flutter/material.dart';
+
+class DotIndicator extends StatelessWidget {
+  const DotIndicator({
+    Key? key,
+    required this.carouselItems,
+    required CarouselController controller,
+    required int current,
+  })  : _controller = controller,
+        _current = current,
+        super(key: key);
+
+  final List<Widget> carouselItems;
+  final CarouselController _controller;
+  final int _current;
+
+  @override
+  Widget build(BuildContext context) {
+    return Row(
+      mainAxisAlignment: MainAxisAlignment.center,
+      children: carouselItems.asMap().entries.map((entry) {
+        return GestureDetector(
+          onTap: () => _controller.animateToPage(entry.key),
+          child: Container(
+            width: 12.0,
+            height: 12.0,
+            margin: EdgeInsets.symmetric(vertical: 8.0, horizontal: 4.0),
+            decoration: BoxDecoration(
+                shape: BoxShape.circle,
+                color: Colors.white
+                    .withOpacity(_current == entry.key ? 0.9 : 0.4)),
+          ),
+        );
+      }).toList(),
+    );
+  }
+}
